+++ Pandora/Pattern/Transformer/Hoistable.hs
@@ -11,3 +11,7 @@ import Pandora.Pattern.Functor.Covariant (Covariant)
 
 class Hoistable t where
 	hoist :: Covariant u => u ~> v -> t u ~> t v
+
+-- data family Schematron (c :: (* -> *) -> k) (t :: * -> *) (u :: * -> *) :: * -> *
+--
+-- instance Hoistable (Schematron c t) where