+++ CHANGELOG.md
@@ -265,3 +265,4 @@
 * Rewrite `zoom` function with right adjunction and move it to `Inventory` module
 * Move `Adjoint` instance for `(:*:)` and `(->)` to `Paradigm.Primary.Functor` module
 * Remove `Schematic` module and move `Schematic` type family to `Interpreted` module
+* Change arity of `Schematic` type family